lambdas/seo-enricher/requirements.txt
google-ads>=24.0.0
requests>=2.28.0
